pyproject.toml
[tool.poetry]
name = "vims-code"
version = "0.1.0b"
description = ""
authors = ["semak <semakmillev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
psycopg2-binary = "^2.9.1"
APScheduler = "^3.8.0"
gevent = "^21.8.0"
configparser = "^5.0.2"
google-api-python-client = "^2.25.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
PyYAML = "^5.4.1"
SQLAlchemy = "^1.4.25"
aiopg = "^1.3.2"
asyncpg = "^0.24.0"
aiofile = "^3.7.2"
aiohttp-middlewares = "^1.1.0"
poetry = "^1.1.11"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
run_server = "vims_code.__main__:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
